requirements.txt
# Database Drivers
psycopg2-binary   # PostgreSQL driver
sqlalchemy        # ORM for database operations
alembic           # Schema migrations
redis
pymongo
python-dotenv
pandas
scikit-learn
joblib
numpy
fastapi
uvicorn
pydantic-settings
xgboost
flaml==2.3.0
mapie==0.8.6

asyncua==1.1.0
python-snap7==1.3
paho-mqtt==2.1.0
watchdog==4.0.0
# Testing & Auditing Tools
ruff              # Linting and logic errors
bandit            # Security vulnerability scanning
locust            # Load testing
pytest            # Unit testing
httpx             # Async testing
requests          # HTTP downloads (FEMTO, NASA data)
kaggle            # Kaggle API (download_kaggle_data.py)

# Authentication & Security
python-jose[cryptography]  # JWT token handling
passlib[bcrypt]            # Password hashing with bcrypt
python-multipart           # Form data handling for OAuth2
slowapi                    # Rate limiting for FastAPI
scipy==1.13.0
prefect==3.1.0
flwr==1.11.0
